feat: transforma countClick de controller pra middleware
